refactor slate links to withLinks
